User-finder for GitHub users
This project is proposed on Brad Traversy's course Modern Javascript - via Udemy.
It uses the API provided by GitHub and served as a course for ES6+ features like, classes, fetch and async/await.
I made a few changes to the UI regardind the alert show/hide behaviour and made sure the website/blog link (if any!) is functionally displayed.